pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rta"
version = "0.1.0"
description = "Run-based Research Thinking Agent pipeline"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "google-genai>=0.5.0",
  "python-dotenv>=1.0.1"
]

[project.scripts]
rta = "rta.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
